backend/src/combatpower/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
redis==5.0.1
ratelimit==2.2.1
pydantic==2.5.0
numpy==1.26.2
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
pillow==10.1.0

